Begin implementing generic traits
Still need to parse and store the generic parameters for `use` and
figure out how to apply the generic type parameter to things such
as `T::class`, `new T()`, etc.

Changing the signature should not be too difficult. We have to copy
the method entry when using a trait anyway so adjust the types there.