Add dictionary test case for parameterized trait
This is the beginning of a test case that is more rigorous, checking
that the trait actually works. It should be more complex in the
future by asserting that certain things fail as well as checking the
iterator.

However this is already causing a segfault so it was committed as-is.