Add Spl\ForwardArrayIterator
ForwardArrayIterator is a simple array iterator which does not dup
the array (though if the user modifies it after making an iterator
then normal copy-on-write will apply and make a copy then).
The ForwardArrayIterator API is intentionally very small.

The name is partly in anticipation of Spl\ReverseArrayIterator,
which will hopefully be the next commit.