test: allow other zend extensions to not fail the test
We (Datadog) run some of the phpt suite with the tracer loaded to be
sure that tests don't crash or are otherwise returning bad results.

Sometimes we skip tests, but this one looks interesting from a
stability perspective. If nobody minds, I'd like to push this patch
upstream rather than maintain it downstream.