Add Spl\ForwardArrayIterator and Spl\ReverseArrayIterator
ForwardArrayIterator is a simple array iterator which does not dup
the array (though if the user modifies it after making an iterator
then normal copy-on-write will apply and make a copy then), and iterates
in normal order. The ReverseArrayIterator is similar but iterates in
reverse order
The APIs are intentionally very small for now.